feat: add stories TextField
