--- src/components/ui/text-field/text-field.stories.ts
@@ -1,30 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react'
-
-import { TextField } from './'
-
-const meta = {
-  title: 'Components/TextField',
-  component: TextField,
-  tags: ['autodocs'],
-  argTypes: {
-    type: {
-      options: ['text', 'password', 'search'],
-      control: { type: 'radio' },
-    },
-    label: {
-      options: ['text', 'password', 'search'],
-      control: { type: 'radio' },
-    },
-  },
-} satisfies Meta<typeof TextField>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const Text: Story = {
-  args: {
-    type: 'text',
-    label: 'text',
-    value: 'Email',
-  },
-}
